LocalRenameJob: Use a DB lock instead of checking status=='inprogress'
It is possible for a rename job to set status to 'inprogress', then
be shut down in such a way that the teardown callback is bypassed
and the status is never set to 'failed', for example if the machine
running the job is unplugged (or a Kubernetes pod is terminated).

When this happens, the job runner will retry the job, but our own
code still sees status=='inprogress' and thinks the job is already
running. Change this check to a database advisory lock, which should
be released if the job disappears for any reason.

I am adding this on top of the other existing checks instead of
replacing them out of abundance of caution.

Bug: T402830
Change-Id: I67a25d868b6976fe02660dc0f7b7bcdd11f60d7a