tests: remove MobileFrontend.Context
SharedDomainUtilsTest and SsoHookHandlerTest craft a SharedDomainUtils
which is created with the service `MobileFrontend.Context`.  This thus
requires MobileFrontend to be enabled.

On CI, we would like to remove the recursive injection of extensions
dependencies. When CentralAuth is a dependency, we are then required to
add MobileFrontend because of this test.  However the test does not seem
to make any assertion related to the mobile context and if I remove the
service, the tests are still passing.

Remove the `MobileFrontend.Context` service getter from these tests.

Backport note: the original patched a mock in
SharedDomainHookHandlerTest instead.

Bug: T389998
Change-Id: Icc71dfe81354cd975c159c85f3cea4967db49b47
(cherry picked from commit 684f8fcbe2574181aae18aa2a04a269360cf959f)