tests: remove MobileFrontend.Context
The ShareDomainHookHandler tests craft a mock SharedDomainUtils which is
created with the service `MobileFrontend.Context`.  This thus requires
MobileFrontend to be enabled.

On CI, we would like to remove the recursive injection of extensions
dependencies. When CentralAuth is a dependency, we are then required to
add MobileFrontend because of this test.  However the test does not seem
to make any assertion related to the mobile context and if I remove the
service, the tests are still passing.

Remove the `MobileFrontend.Context` service getter from the mock.

I ran tests with the command ./phpunit ../extensions/CentralAuth with
and without MobileFrontend.

Bug: T389998
Change-Id: Icc71dfe81354cd975c159c85f3cea4967db49b47