Use the `$user` object when performer is passed in as NULL
Instead of callers to pass in the user object twice (which is not
very readable), the autoCreateUser() method can instead substitute
the $performer with the user object if NULL is passed in.

The method has been documented to backup this claim and suggests
to consumers of this code what to do in either case.

Bug: T408724
Depends-On: If43bcc4c1f0da80282ac74fb0eaa956929714324
Change-Id: I16497e4789d38d1ade07aceeebb4031359bd4b8c