Replace unused Status return value with trivial bool
The status and the value (which is for some reason the string "bad")
is not used anywhere for anything. The method can as well return a
bool.

I checked and the protected method is not used anywhere outside of
this class. I will mark it as private in the next patch.

Change-Id: I3ad49d729f319272e56f4b13aa78ff4c02a642e9