SharedDomainHookHandler: Don't loop PasswordReset via login flow
Removing PasswordReset from central special pages would not be ideal.
We want to still do password reset on the central domain whether or
not the user is logged-in.

In this patch, we introduced another restriction type (special pages
that may not required authentication to access them). In addition to
the ones that require the user to be authenticated, we can have a
list of pages that may or may not require authentication. An example
is the PasswordReset special page.

NOTE: Cleanup some unused class properties.

Bug: T409984
Change-Id: Ib9a936cbcb4e9f19b3f9ad58fce01104af515c15