CentralAuthUser: Add $recency param to queryAttached(), queryUnattached()
Many callers of these methods do not need the current behavior of
querying the primary database for each wiki if its replica database
doesn't have the expected row. They would be fine with just querying
the replica. It's hard to review them all, though.

Add a parameter to control it. The default remains set to
READ_LATEST_IMMUTABLE to avoid changes in behavior. Callers may
use READ_NORMAL to avoid warnings about database inconsistencies
(which are to be expected when reading from replicas).

Use READ_NORMAL in CentralAuthUser::getLocalGroups(), which has been
causing the log spam in T385310.

Bug: T385310
Change-Id: Icac17ea1de01ed18bdc5989f14b2e78153b8965c