SUL3: Show expiry message only if user isn't logged-in centrally
... but logged-in locally.

This warning message per I9f38ab6ef46f6a0854d9 (via a URL param)
is supposed to be shown if the user doesn't have a central session
active but has a local session active (logged-in locally). I think
I missed this during testing the patch locally.

The patch now hides the message when the user is logged-in on the
shared domain which wasn't the case before. Note that the warning
URL param is still sent from the local domain to the shared domain
but now when we `onLoginFormValidErrorMessages()` hook handler, it
doesn't add the message on the shared domain, hence not shown when
the user is logged-in.

Follow-up on I9f38ab6ef46f6a0854d9cc5f5abeda12c0b5a777.

Test plan
=========

1. Login normally (both central and local session active)
2. Head to the shared domain and delete cookies related to the
   local domain on the shared domain (fake central session expiry).
3. Head back to the local domain and verify you're still logged-in
   locally and then visit `Special:CreateAccount` or `...:Login`
   and you'll be redirected to the shared domain (as a logged out
   user).
3. The central session expiry message should be shown if you're
   logged-out on the shared domain. Before this patch, it shows
   even when you're logged in.

Bug: T411822
Bug: T393459
Change-Id: I37ae7dca3e32a9308c8c53f32df0ede7a4bad07f