Add localuser.lu_attachment_method field patch to schema updater
The patch was added in 2022 (c425e72a18fa8054e401), but it could not
be added to the automatic schema updater until the migration to
virtual domains in 2024 (8326b0acda7deaedc864).

We recently started writing to that field (0149741a458fb85cbd83),
breaking all CentralAuth installations created before that patch
which did not manually apply the schema update.

Note that you need to run `update.php --doshared` to apply schema
updates affecting virtual domains.

Bug: T304015
Change-Id: I7a36c4ffe389d1dc85b40442be360fb1516682eb