ForeignRest: Always use 'centralauthtoken' for auth'd requests
We had attempted an optimization where, if the user was also logged
in on the foreign wiki, we skipped the 'centralauthtoken' fetching
and relied on the browser credentials to authenticate the requests.
There is a similar optimization in ForeignApi.

It turns out that, unlike the action API, the REST API requires a
config option to allow doing that ($wgRestAllowCrossOriginCookieAuth),
and enabling that option has complex performance/caching effects that
mean it may never be enabled on Wikimedia wikis. The optimization
doesn't work without that option, causing all requests to fail.

Remove the code, making ForeignRest work correctly on Wikimedia wikis.
Since the 'centralauthtoken' mode always works, we can always use it,
gaining reliability and simplicity at the cost of some efficiency.

Bug: T412530
Change-Id: I08319d90faf5897f1e7b8bf4dc055391f593a562